move consolidated elements to ChannelPageSkeleton
