+++ frontends/mit-open/src/pages/ChannelPage/UnitChannelSkeleton.tsx
@@ -13,7 +13,7 @@ import { SourceTypeEnum } from "api"
 import { getSearchParamMap } from "@/common/utils"
 import { HOME as HOME_URL, UNITS as UNITS_URL } from "../../common/urls"
 import { ChannelTypeEnum } from "api/v0"
-import { UNITS_LABEL } from "./ChannelPage"
+import { UNITS_LABEL } from "./ChannelPageSkeleton"
 
 export const ChannelTitleRow = styled.div`
   display: flex;
